Allow (most) debugging without rebuilding nwjs app
